refactor: 리액트 쿼리 커스텀 훅과 apis에 중복해서 있던 api요청 함수 제거
