+++ src/hooks/useUpdateTask.ts
@@ -1,54 +1,7 @@
 import useToast from './useToast';
 
-import { api } from '@apis';
+import { createNewTask, updateTask, deleteTask, dragTask, TaskInfo, UpdateTaskInfo, DragInfo } from '@apis/index';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-
-interface TaskInfo {
-  planId: number;
-  tabId: number;
-  assigneeId: number | undefined;
-  title: string;
-  description: string;
-  startDate: string | null;
-  endDate: string | null;
-  labels: number[];
-}
-interface UpdateTaskInfo extends TaskInfo {
-  taskId: number;
-}
-
-interface DragInfo {
-  planId: number;
-  targetTabId: number;
-  targetId: number;
-  newPrevId: number | null;
-}
-
-export const createNewTask = async (params: TaskInfo): Promise<void> => {
-  await api.post('/api/tasks', params);
-};
-
-export const updateTask = async (params: UpdateTaskInfo) => {
-  await api.put(`/api/tasks/${params.taskId}`, {
-    planId: params.planId,
-    tabId: params.tabId,
-    assigneeId: params.assigneeId,
-    title: params.title,
-    description: params.description,
-    startDate: params.startDate,
-    endDate: params.endDate,
-    labels: params.labels,
-  });
-};
-
-export const deleteTask = async (params: { taskId: number }) => {
-  await api.delete(`/api/tasks/${params.taskId}`);
-};
-
-export const dragTask = async (params: DragInfo) => {
-  const response = await api.put('api/tasks/change-order', params);
-  return response;
-};
 
 export function useUpdateTask(planId: number) {
   const queryClient = useQueryClient();
